Handle IOException from Response
The type system still thinks it could through an IOException, but a
generic 500 response uses btye array streams which don't really throw
IOExceptions even though they declare that they do.

I also changed parser exception handlers to return 400 instead of 500.
If these exceptions are thrown it means the request could not be parsed
because it is not a valid HTTP message.