Fix bug causing 400 responses on mac
This turned out to be a bug in the HTTP specification.  The way the ABNF
is written, it breaks on space separated header field strings if one of
those strings is one character long.  It was breaking when trying to
parse the `User-Agent` header which included the string "Mac OS X 10_".
It failed to parse past the "X".